requirements.txt
watchdog==4.0.1
pyyaml==6.0.2
send2trash==1.8.3
numpy==2.4.1
faster-whisper==1.2.1
ctranslate2==4.6.3
opencv-python==4.12.0.88